implemented some more methods and pulled changes from the master branch
