+++ DAO-Pattern/nbproject/private/configs/Run_in_Browser.properties
@@ -1,0 +1,4 @@
+# Do not modify this property in this configuration. It can be re-generated.
+javafx.run.as=embedded
+javafx.run.inbrowser=<Default System Browser>
+javafx.run.inbrowser.path=C:\\Program Files (x86)\\Mozilla Firefox\\firefox.exe
