+++ DAO-Pattern/nbproject/private/private.properties
@@ -1,0 +1,7 @@
+auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
+compile.on.save=true
+do.depend=false
+do.jar=true
+javac.debug=true
+javadoc.preview=true
+user.properties.file=C:\\Users\\Hendrik\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
