+++ DAO-Pattern/nbproject/project.properties
@@ -1,6 +1,5 @@
 annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
-annotation.processing.processor.options=
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
@@ -30,15 +29,19 @@ dist.jar=${dist.dir}/DAO-Pattern.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
+file.reference.sqlite-jdbc-3.20.0.jar=src/dataAccessObjects/sqlite-jdbc-3.20.0.jar
 includes=**
 # Non-JavaFX jar file creation is deactivated in JavaFX 2.0+ projects
 jar.archive.disabled=true
 jar.compress=false
 javac.classpath=\
-    ${javafx.classpath.extension}
+    ${javafx.classpath.extension}:\
+    ${file.reference.sqlite-jdbc-3.20.0.jar}:\
+    ${libs.SQLite.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
+javac.external.vm=false
 javac.processorpath=\
     ${javac.classpath}
 javac.source=1.8
@@ -100,6 +103,8 @@ manifest.custom.codebase=*
 manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
+mkdist.disabled=false
+no.dependencies=true
 platform.active=default_platform
 run.classpath=\
     ${dist.jar}:\
