+++ DAO-Pattern/src/dataAccessObjects/DataLayerSqlite.java
@@ -1,0 +1,16 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package dataAccessObjects;
+
+/**
+ *
+ * @author Hendrik
+ */
+public class DataLayerSqlite implements IDataLayer {
+    public ITrainerDao getTrainerDao() {
+        
+    }
+}
